renamed authentication to auth
