+++ src/app/app-routing.module.ts
@@ -11,8 +11,7 @@ const routes: Routes = [
       { path: 'home', component: HomeComponent },
       {
         path: 'auth',
-        loadChildren:
-          '@app/authentication/authentication.module#AuthenticationModule'
+        loadChildren: '@app/auth/auth.module#AuthModule'
       },
       { path: 'cart', loadChildren: '@app/cart/cart.module#CartModule' },
       {
