+++ src/app/auth/acknowledgement/acknowledgement-dialog.component.scss
