+++ src/app/auth/acknowledgement/acknowledgement-dialog.component.spec.ts
