+++ src/app/auth/auth-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AcknowledgementDialogComponent } from '@app/authentication/acknowledgement/acknowledgement-dialog.component';
-import { ForgotPasswordDialogComponent } from '@app/authentication/forgot-password/forgot-password-dialog.component';
-import { LoginDialogComponent } from '@app/authentication/login/login-dialog.component';
-import { LogoutComponent } from '@app/authentication/logout/logout.component';
-import { SignupComponent } from '@app/authentication/signup/signup.component';
+import { AcknowledgementDialogComponent } from '@app/auth/acknowledgement/acknowledgement-dialog.component';
+import { ForgotPasswordDialogComponent } from '@app/auth/forgot-password/forgot-password-dialog.component';
+import { LoginDialogComponent } from '@app/auth/login/login-dialog.component';
+import { LogoutComponent } from '@app/auth/logout/logout.component';
+import { SignupComponent } from '@app/auth/signup/signup.component';
 
 const routes: Routes = [
   { path: '', component: LoginDialogComponent },
