+++ src/app/auth/auth.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
-import { AuthenticationRoutingModule } from '@app/authentication/authentication-routing.module';
+import { AuthenticationRoutingModule } from '@app/auth/auth-routing.module';
 import { SharedModule } from '@shared/shared.module';
 
 @NgModule({
   imports: [SharedModule, AuthenticationRoutingModule],
   declarations: [AuthenticationRoutingModule.components]
 })
-export class AuthenticationModule {}
+export class AuthModule {}
