+++ src/app/auth/forgot-password/forgot-password-dialog.component.scss
