+++ src/app/auth/forgot-password/forgot-password-dialog.component.spec.ts
