+++ src/app/auth/guards/authentication.guard.spec.ts
@@ -1,0 +1,14 @@
+import { inject, TestBed } from '@angular/core/testing';
+import { AuthGuard } from './auth.guard';
+
+describe('AuthenticationGuard', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [AuthGuard]
+    });
+  });
+
+  it('should ...', inject([AuthGuard], (guard: AuthGuard) => {
+    expect(guard).toBeTruthy();
+  }));
+});
