+++ src/app/auth/login/login-dialog.component.scss
