+++ src/app/auth/models/dialog.interface.ts
