+++ src/app/auth/models/user.model.ts
