+++ src/app/auth/services/auth.service.spec.ts
@@ -1,6 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-
-import { AuthenticationService } from './authentication.service';
+import { AuthenticationService } from './auth.service';
 
 describe('AuthenticationService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
