+++ src/app/auth/services/dialog.interface.ts
