--- src/app/authentication/guards/authentication.guard.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, async, inject } from '@angular/core/testing';
-
-import { AuthenticationGuard } from './authentication.guard';
-
-describe('AuthenticationGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AuthenticationGuard]
-    });
-  });
-
-  it('should ...', inject([AuthenticationGuard], (guard: AuthenticationGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
